Added search controller
