+++ src/http/controllers/core.controller.ts
@@ -1,12 +1,9 @@
 import {
     BadRequestException,
-    CacheInterceptor,
     Controller,
     Get,
     Logger,
-    NotFoundException,
     Param,
-    UseInterceptors
 } from '@nestjs/common';
 import {InjectQueue} from '@nestjs/bull';
 import {ConfigService} from "@nestjs/config";
@@ -15,8 +12,6 @@ import {MessagePattern, Payload} from '@nestjs/microservices';
 import {plainToClass} from 'class-transformer';
 
 import {Queue} from 'bull';
-
-import {LumConstants} from '@lum-network/sdk-javascript';
 
 import {LumService, LumNetworkService, BlockService, TransactionService} from '@app/services';
 import {QueueJobs, Queues} from '@app/utils';
@@ -37,29 +32,6 @@ export class CoreController {
         private readonly _transactionService: TransactionService
     ) {
     }
-
-    @Get('search/:data')
-    @UseInterceptors(CacheInterceptor)
-    async search(@Param('data') data: string) {
-        // We check the different combinations
-        if (/^\d+$/.test(data)) {
-            return {type: 'block', data};
-        } else if (String(data).startsWith(LumConstants.LumBech32PrefixValAddr)) {
-            return {type: 'validator', data};
-        } else if (String(data).startsWith(LumConstants.LumBech32PrefixAccAddr)) {
-            return {type: 'account', data};
-        } else {
-            if (await this._blockService.get(parseInt(data, 10))) {
-                return {type: 'block', data};
-            } else if (await this._transactionService.get(data)) {
-                return {type: 'transaction', data};
-            } else {
-                throw new NotFoundException('data_not_found');
-            }
-        }
-    }
-
-
 
     @Get('lum')
     async lum() {
